+++ cpp/src/img/index.cpp
@@ -40,8 +40,8 @@ set<string> Index::getPossibleWords() {
     return words;
 }
 
-map<string, double> Index::getScores(const Image &image) {
-    map<string, double> scores;
+map<string, float> Index::getScores(const Image &image) {
+    map<string, float> scores;
     vector<string> words_vect = image.getWords();
     set<string> words(words_vect.begin(), words_vect.end());
     for(auto& word : words) {
