+++ cpp/src/img/index.h
@@ -16,7 +16,7 @@ class Index {
 public:
     void indexImage(Image &image);
     void indexImages(vector<Image>& images);
-    map<string, double> getScores(const Image &image);
+    map<string, float> getScores(const Image &image);
     set<string> getImagesPathWithWord(string word);
     set<string> getPossibleWords();
 private:
